requirements.txt
mysql-connector-python
pandas
wxPython